requirements.txt
azure-common==1.1.28
azure-core==1.30.0
azure-functions==1.18.0
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-web==7.2.0
azure-search-documents==11.6.0b1
azure-storage-blob==12.19.0
llama-index==0.10.6
llama-index-agent-openai==0.1.1
llama-index-core==0.10.6.post1
llama-index-embeddings-azure-openai==0.1.1
llama-index-embeddings-openai==0.1.1
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.1
llama-index-llms-openai==0.1.2
llama-index-multi-modal-llms-openai==0.1.1
llama-index-program-openai==0.1.2
llama-index-question-gen-openai==0.1.1
llama-index-readers-file==0.1.3
llama-index-readers-web==0.1.5
llama-index-vector-stores-azureaisearch==0.1.3
llamaindex-py-client==0.1.12
marshmallow==3.20.2
matplotlib==3.8.3
numpy==1.26.4
openai==1.12.0
flask_cors
flask




